refactor: folders architecture
